Use the proper C++ function

+++ DESIRE-Engine/src/Core/math/math.h
@@ -99,7 +99,7 @@ inline float Clamp(float x, float min, float max)
 	_mm_store_ss(&x, _mm_max_ss(_mm_min_ss(_mm_load_ss(&x), _mm_load_ss(&max)), _mm_load_ss(&min)));
 	return x;
 #else
-	return std::fmaxf(std::fminf(x, max), min);
+	return std::fmax(std::fmin(x, max), min);
 #endif
 }
